requirements.txt
einops
flax
absl-py

# For T5 tokenizer
transformers
sentencepiece

# Just used for image resizing
torch
torchvision

Pillow==9.4.0
numpy